--- a/quic/s2n-quic-core/src/recovery/cubic.rs
+++ b/quic/s2n-quic-core/src/recovery/cubic.rs
@@ -193,6 +193,7 @@ impl CongestionController for CubicCongestionController {
         &mut self,
         time_sent: Timestamp,
         bytes_sent: usize,
+        app_limited: Option<bool>,
         rtt_estimator: &RttEstimator,
     ) {
         if bytes_sent == 0 {
@@ -204,7 +205,9 @@ impl CongestionController for CubicCongestionController {
             .try_add(bytes_sent)
             .expect("bytes sent should not exceed u32::MAX");
 
-        self.under_utilized = self.is_congestion_window_under_utilized();
+        if let Some(app_limited) = app_limited {
+            self.under_utilized = app_limited && self.is_congestion_window_under_utilized();
+        }
 
         if let Recovery(recovery_start_time, RequiresTransmission) = self.state {
             // A packet has been sent since we entered recovery (fast retransmission)