--- a/quic/s2n-quic-core/src/recovery/cubic/tests.rs
+++ b/quic/s2n-quic-core/src/recovery/cubic/tests.rs
@@ -216,7 +216,7 @@ fn on_packet_sent() {
     cc.congestion_window = 100_000.0;
 
     // Last sent packet time updated to t10
-    cc.on_packet_sent(now + Duration::from_secs(10), 1, &rtt_estimator);
+    cc.on_packet_sent(now + Duration::from_secs(10), 1, None, &rtt_estimator);
 
     assert_eq!(cc.bytes_in_flight, 1);
 
@@ -250,7 +250,7 @@ fn on_packet_sent() {
     );
 
     // Last sent packet time updated to t20
-    cc.on_packet_sent(now + Duration::from_secs(20), 1, &rtt_estimator);
+    cc.on_packet_sent(now + Duration::from_secs(20), 1, None, &rtt_estimator);
 
     assert_eq!(cc.bytes_in_flight, 2);
 
@@ -273,7 +273,7 @@ fn on_packet_sent_application_limited() {
     cc.state = SlowStart;
 
     // t0: Send a packet in Slow Start
-    cc.on_packet_sent(now, 1000, &rtt_estimator);
+    cc.on_packet_sent(now, 1000, Some(true), &rtt_estimator);
 
     assert_eq!(cc.bytes_in_flight, 93_500);
     assert_eq!(cc.time_of_last_sent_packet, Some(now));
@@ -284,7 +284,12 @@ fn on_packet_sent_application_limited() {
     assert!(!cc.under_utilized);
 
     // t15: Send a packet in Congestion Avoidance
-    cc.on_packet_sent(now + Duration::from_secs(15), 1000, &rtt_estimator);
+    cc.on_packet_sent(
+        now + Duration::from_secs(15),
+        1000,
+        Some(true),
+        &rtt_estimator,
+    );
 
     assert_eq!(cc.bytes_in_flight, 94_500);
     assert_eq!(
@@ -295,7 +300,12 @@ fn on_packet_sent_application_limited() {
 
     // t20: Send packets to fully utilize the congestion window
     while cc.bytes_in_flight < cc.congestion_window() {
-        cc.on_packet_sent(now + Duration::from_secs(20), 1000, &rtt_estimator);
+        cc.on_packet_sent(
+            now + Duration::from_secs(20),
+            1000,
+            Some(true),
+            &rtt_estimator,
+        );
     }
 
     assert!(!cc.under_utilized);
@@ -311,7 +321,7 @@ fn on_packet_sent_fast_retransmission() {
     cc.bytes_in_flight = BytesInFlight::new(99900);
     cc.state = Recovery(now, RequiresTransmission);
 
-    cc.on_packet_sent(now + Duration::from_secs(10), 100, &rtt_estimator);
+    cc.on_packet_sent(now + Duration::from_secs(10), 100, None, &rtt_estimator);
 
     assert_eq!(cc.state, Recovery(now, Idle));
 }
@@ -334,7 +344,7 @@ fn congestion_avoidance_after_idle_period() {
     cc.state = SlowStart;
 
     // t0: Send a packet in Slow Start
-    cc.on_packet_sent(now, 1000, rtt_estimator);
+    cc.on_packet_sent(now, 1000, Some(true), rtt_estimator);
 
     assert_eq!(cc.bytes_in_flight, 1000);
 
@@ -343,7 +353,12 @@ fn congestion_avoidance_after_idle_period() {
     cc.state = State::congestion_avoidance(now + Duration::from_secs(10));
 
     // t15: Send a packet in Congestion Avoidance while under utilized
-    cc.on_packet_sent(now + Duration::from_secs(15), 1000, rtt_estimator);
+    cc.on_packet_sent(
+        now + Duration::from_secs(15),
+        1000,
+        Some(true),
+        rtt_estimator,
+    );
     assert!(cc.is_congestion_window_under_utilized());
 
     assert_eq!(cc.bytes_in_flight, 2000);
@@ -371,7 +386,12 @@ fn congestion_avoidance_after_idle_period() {
 
     // t20: Send packets to fully utilize the congestion window
     while cc.bytes_in_flight < cc.congestion_window() {
-        cc.on_packet_sent(now + Duration::from_secs(20), 1000, rtt_estimator);
+        cc.on_packet_sent(
+            now + Duration::from_secs(20),
+            1000,
+            Some(false),
+            rtt_estimator,
+        );
     }
 
     assert!(!cc.is_congestion_window_under_utilized());
@@ -729,7 +749,7 @@ fn on_packet_ack_utilized_then_under_utilized() {
     cc.congestion_window = 100_000.0;
     cc.state = SlowStart;
 
-    cc.on_packet_sent(now, 60_000, &rtt_estimator);
+    cc.on_packet_sent(now, 60_000, Some(true), &rtt_estimator);
     cc.on_ack(now, 50_000, (), &rtt_estimator, random, now);
     let cwnd = cc.congestion_window();
 
@@ -752,7 +772,7 @@ fn on_packet_ack_utilized_then_under_utilized() {
 
     // Now the application has had a chance to send more data, but it didn't send enough to
     // utilize the congestion window, so the window does not grow.
-    cc.on_packet_sent(now, 1200, &rtt_estimator);
+    cc.on_packet_sent(now, 1200, Some(true), &rtt_estimator);
     assert!(cc.under_utilized);
     cc.on_ack(
         now,